requirements.txt
# Core & config
pydantic>=2.7.0
pydantic-settings>=2.2.0
orjson>=3.10.0

# LLM & routing
openai>=1.55.0
tiktoken>=0.7.0
tenacity>=9.0.0

# PDF parsing (text + tables)
PyMuPDF>=1.24.9           # pymupdf
pdfplumber>=0.11.0
camelot-py[cv]>=0.11.0
opencv-python-headless>=4.10.0.84
# Optional alternative backend (Java required)
# tabula-py>=2.9.0

# NLP/regex & fuzzy
regex>=2024.5.15
rapidfuzz>=3.9.0

# Storage & retrieval
duckdb>=1.0.0
pandas>=2.2.0
sqlglot>=25.0.0
# Optional semantic retrieval
# faiss-cpu>=1.8.0
# sentence-transformers>=3.0.0

# CLI chatbot UX
typer>=0.12.5
rich>=13.8.0
prompt-toolkit>=3.0.47
click-spinner>=0.1.10

# Observability & ops
structlog>=24.4.0
sentry-sdk>=2.12.0
diskcache>=5.6.3

# Time & units
pendulum>=3.0.0
pint>=0.24.0

# FastAPI service (HTTP + streaming)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0

# Scale (optional; useful for 1000+ PDFs)
ray>=2.36.0
aiofiles>=24.1.0

# Dev & testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
hypothesis>=6.112.0
coverage>=7.6.0
ruff>=0.6.9
mypy>=1.10.0
types-python-dateutil

faiss-cpu>=1.8.0
numpy>=1.26.0
pyarrow>=15.0.0